Add pagination and result orderer tests to own file. Update service tests
