+++ browserTests/views/serviceTest.js
@@ -1,7 +1,8 @@
 import { waitForReact, ReactSelector } from 'testcafe-react-selectors';
 import { Selector } from 'testcafe';
 import config from '../config';
-import { getLocation } from '../utility';
+import paginationTest from '../utility/paginationTest';
+import resultOrdererTest from '../utility/resultOrdererTest';
 
 const { server } = config;
 
@@ -31,127 +32,23 @@ fixture `Service page tests`
     await waitForReact();
   });
 
-test('Keyboard navigation is OK', async (t) => {
-  const input = Selector('#SearchBar input');
-  let select = Selector('#result-sorter')
-  const listItems = Selector('#paginatedList-events li[role="link"]')
 
-  const alphaDescText = await listItems.nth(0).textContent;
+paginationTest(servicePage);
 
-  // Use keyboard to change result ordering
-  await t
-    .click(input)
-    .pressKey('tab') // Tabs to search icon button
-    .pressKey('tab') // Result orderer
-    // .expect(select.getReact(({props}) => props.value)).eql('match-desc')
-    // .pressKey('down')
-    .pressKey('down');
+resultOrdererTest();
 
-  const alphaAscText = await listItems.nth(0).textContent;
-  await t
-    .expect(alphaAscText).notEql(alphaDescText)
-    .pressKey('down');
+test('Service view list item click takes to correct unit view', async (t) => {
+  const units =  Selector('#paginatedList-events li[role="link"]');
+  const name = await units.nth(0).find('p[role="textbox"]').textContent;
+  const unitTitleSelector = Selector('.TitleText');
+  const backToServiceButton = Selector('#SearchBar .SMBackButton');
+  const servicePageTitle = Selector('#view-title');
 
-  const accessibilityDescText = await listItems.nth(0).textContent;
   await t
-    .expect(accessibilityDescText).notEql(alphaAscText)
-    .pressKey('up');
-  
-  const accessibilityAscText = await listItems.nth(0).textContent;
-  await t
-    .expect(accessibilityAscText).notEql(accessibilityDescText)
-  ;
-
-  // Use keyboard to navigate result list
-  await t
-    .pressKey('tab')
-    .expect(listItems.nth(0).focused).ok()
-    .pressKey('tab')
-    .expect(listItems.nth(1).focused).ok()
-    .pressKey('shift+tab')
-    .expect(listItems.nth(0).focused).ok()
-  ;
-
-  const pagination = ReactSelector('PaginationComponent');
-  const buttons = pagination.find('button');
-  const location = getLocation();
-  await t
-    // Click next page button
-    .click(buttons.nth(1))
-    // Focus is lost to start of the page after button click
-    // Pagination moves focus to p element with tabindex -1 which doesn't seem to work
-    // with testCafe tests but works on live version
-    .expect(location).contains(servicePage)
-    .expect(location).contains('p=2')
-    .click(select) // Click select to move focus back to view
-  ;
-  
-  // After clicking to change page focus is moved to start of the list
-  // We need to move back to page navigation
-  for (let i = 0; i < 12; i++) {
-    await t.pressKey('tab');
-  }
-
-  // Check keyboard navigation
-  await t
-    .pressKey('shift+tab') // Move back to previous page button
-    .expect(buttons.nth(0).focused).ok()
-    .pressKey('tab') // next page button
-    .expect(buttons.nth(1).focused).ok()
-    .pressKey('tab') // 1st page button
-    .expect(buttons.nth(2).focused).ok()
-    // Tab to 3rd page element since 2nd is active and tabindex -1
-    .pressKey('tab') // 3 page button
-    .expect(buttons.nth(4).focused).ok()
-    .pressKey('enter')
-    .expect(location).contains('p=3')
+    .click(units.nth(0))
+    .expect(unitTitleSelector.textContent).eql(name)
+    .click(backToServiceButton)
+    .expect(servicePageTitle.focused).ok('Expect service page title to be focused')
   ;
 });
 
-test('Pagination attributes change correctly', async (t) => {
-  const pagination = ReactSelector('PaginationComponent');
-  const buttons = pagination.find('button');
-  const previousPageButton = buttons.nth(0);
-  const nextPageButton = buttons.nth(1);
-  await t
-    .expect(previousPageButton.getAttribute('tabindex')).eql('-1')
-    .expect(previousPageButton.getAttribute('disabled')).eql('')
-    .click(nextPageButton)
-    .expect(previousPageButton.getAttribute('disabled')).notOk()
-    .expect(previousPageButton.getAttribute('tabindex')).eql('0')
-  ;
-
-  const location = getLocation();
-  await t
-    // Expect page id to exists in url
-    .expect(location).contains(servicePage)
-    .expect(location).contains('p=2')
-    // Expect 2nd page element to be set active and have related attributes
-    .expect(buttons.nth(3).getAttribute('disabled')).eql('')
-    .expect(buttons.nth(3).getAttribute('tabindex')).eql('-1')
-    // Click 3rd page element
-    .click(buttons.nth(4))
-    .expect(location).contains('p=3')
-    // Expect 2nd page element to reset attributes
-    .expect(buttons.nth(3).getAttribute('disabled')).notOk()
-    .expect(buttons.nth(3).getAttribute('tabindex')).eql('0')
-    // Expect 3rd page element to be set active and have related attributes
-    .expect(buttons.nth(4).getAttribute('disabled')).eql('')
-    .expect(buttons.nth(4).getAttribute('tabindex')).eql('-1')
-  ;
-});
-
-test('Pagination\'s page change focuses correctly', async(t) => {
-  const resultList = ReactSelector('ResultList');
-  const focusTarget = resultList.find('p').nth(1);
-  const pagination = ReactSelector('PaginationComponent');
-  const buttons = pagination.find('button');
-
-  await t
-    .click(buttons.nth(3))
-    // Focus target should be focused
-    .expect(focusTarget.focused).ok()
-    // Focus target should be screen reader text with pagination information
-    .expect(focusTarget.innerText).contains('sivu 2 kautta')
-  ;
-});
